# docs/05-api/openapi.yaml
...
openapi: 3.1.0
info:
  title: Klick-Pro (servio) HTTP API
  version: 0.0.0-reverse-engineered
  description: >-
    Reverse-engineered from source at commit cd8f4fb (verified 2026-09-16). Source of truth is the implementation; every
    operation carries x-source file:line.


    Conventions and caveats:

    - Responses are raw NextResponse.json; errors are mostly {"error": string}. The apiSuccess/apiError envelope in
    src/lib/api-response.ts is not used.

    - Every POST/PUT/PATCH/DELETE under /api/* must carry an Origin header equal to the app origin or APP_URL, otherwise
    proxy.ts returns 403 {"error":"Request origin is not allowed."}. This applies to webhooks and non-browser clients
    too. Under the custom server only the exact APP_URL origin passes in practice; a genuine same-origin request to a
    different host (e.g. 127.0.0.1 when APP_URL is localhost) also gets 403. [VALIDATED 2026-09-17 · V-20]

    - /api/v1/:path* is rewritten to /api/:path* (next.config.ts, afterFiles). Physical routes /api/v1/messages and
    /api/v1/professionals take precedence and have no /api/* twin. /api/v1/v1/* is not rewritten twice and returns
    404. [VALIDATED 2026-09-17 · V-13]

    - Schemas marked [NEEDS VALIDATION] could not be fully established from code; no properties were invented.

    - Realtime Socket.IO events (path /api/realtime) are documented in docs/05-api/api-specification.md, not here.


    Human-readable reference: docs/05-api/api-specification.md
servers:
...
paths:
...
  /api/profile/locations/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
          minimum: 1
    patch:
      tags:
        - Profile
      operationId: updateSavedLocation
      summary: Update owned saved location
      x-source: app/api/profile/locations/[id]/route.ts:21-44
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - label
                - address
              properties:
                label:
                  type: string
                  minLength: 1
                  maxLength: 80
                address:
                  type: string
                  minLength: 1
                  maxLength: 300
                isPrimary:
                  type: boolean
      responses:
        "200":
          description: Saved location.
          content:
            application/json:
              schema:
                type: object
                required:
                  - location
                properties:
                  location:
                    $ref: "#/components/schemas/ClientSavedLocation"
        "400":
          description: "Error with a human-readable message: {error: string}."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
        "404":
          description: "Error with a human-readable message: {error: string}."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
    delete:
      tags:
        - Profile
      operationId: deleteSavedLocation
      summary: Delete owned saved location (promotes replacement primary)
      x-source: app/api/profile/locations/[id]/route.ts:46-68
      security:
        - cookieAuth: []
      responses:
        "200":
          description: "{success: true}"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessTrue"
        "404":
          description: "Error with a human-readable message: {error: string}."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
...
  /api/professional/jobs/export:
    post:
      tags:
        - Professional
      summary: PDF of the professional's running projects
      operationId: exportProfessionalJobs
      security:
        - cookieAuth: []
      description: Source app/api/professional/jobs/export/route.ts:78
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportRequest"
      responses:
        "200":
          description: PDF attachment running-projects-<scope>.pdf
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        "400":
          description: Invalid export request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Not an active PROFESSIONAL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/professional/proposals:
    get:
      tags:
        - Professional
      summary: The caller's latest proposal for a job and the latest negotiation
      operationId: getOwnProposal
      security:
        - cookieAuth: []
      description: Source app/api/professional/proposals/route.ts:17
      parameters:
        - name: jobId
          in: query
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: Proposal or null
          content:
            application/json:
              schema:
                type: object
                properties:
                  proposal:
                    oneOf:
                      - type: "null"
                      - $ref: "#/components/schemas/ProjectRequestWithActor"
                  negotiation:
                    type:
                      - object
                      - "null"
                    properties:
                      senderRole:
                        type: string
                      previousBidAmount:
                        type:
                          - integer
                          - "null"
                      previousDuration:
                        type:
                          - string
                          - "null"
                      previousMessage:
                        type:
                          - string
                          - "null"
        "400":
          description: A valid job is required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: No cookie
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Not PROFESSIONAL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Unable to load (also invalid JWT)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      tags:
        - Professional
      summary: Create a proposal, or update the existing PENDING one
      operationId: sendProposal
      security:
        - cookieAuth: []
      description: >-
        Source app/api/professional/proposals/route.ts:55. Update returns 200, create returns 201. Notifies the client
        (DB, email, notification:new, proposal:new). Create also queues an admin notification in-process.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - jobId
                - bidAmount
                - duration
                - coverLetter
              properties:
                jobId:
                  type: integer
                  minimum: 1
                bidAmount:
                  type: integer
                  minimum: 1
                  maximum: 10000000
                duration:
                  type: string
                  minLength: 2
                  maxLength: 100
                coverLetter:
                  type: string
                  minLength: 10
                  maxLength: 5000
      responses:
        "200":
          description: Existing pending proposal updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  proposal:
                    $ref: "#/components/schemas/ProjectRequest"
        "201":
          description: Proposal created
          content:
            application/json:
              schema:
                type: object
                properties:
                  proposal:
                    $ref: "#/components/schemas/ProjectRequest"
        "400":
          description: Invalid body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: No cookie
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Not PROFESSIONAL or own job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Job not accepting proposals
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Unable to send
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: servio_session
      description: >-
        HS256 JWT {userId, role, sessionId} issued by src/lib/auth.ts createSession; validated against the sessions
        table on every request. Same cookie for clients, professionals and admins — role is inside the token.
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Same session JWT sent as Authorization: Bearer. Accepted only by the few handlers noted per operation."
  schemas:
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
...
    ReportRequest:
      type: object
      required:
        - scope
      properties:
        scope:
          type: string
          enum:
            - all
            - selected
        ids:
          type: array
          items:
            type: integer
            minimum: 1
        pageSize:
          type: string
          enum:
            - A4
            - LETTER
        orientation:
          type: string
          enum:
            - portrait
            - landscape
      x-source: src/lib/reports/pdf/request.ts:4-9
...
    ProjectRequest:
      type: object
      description: "Prisma ProjectRequest row [NEEDS VALIDATION: full columns]"
      properties:
        id:
          type: integer
        jobId:
          type: integer
        clientId:
          type: integer
        professionalId:
          type: integer
        bidAmount:
          type: integer
        duration:
          type: string
        coverLetter:
          type: string
        status:
          type: string
          description: PENDING | ACCEPTED | REJECTED
        origin:
          type: string
          enum:
            - CLIENT_HIRE
            - PROFESSIONAL_PROPOSAL
        createdAt:
          type: string
          format: date-time
    ProjectRequestWithActor:
      allOf:
        - $ref: "#/components/schemas/ProjectRequest"
        - type: object
          properties:
            lastActorRole:
              type: string
              enum:
                - CLIENT
                - PROFESSIONAL
...
    ErrorMessage:
      type: object
      required:
        - error
      properties:
        error:
          type: string
      x-source: e.g. app/api/auth/[action]/route.ts:44
...
    SuccessTrue:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
          const: true
...
    ClientSavedLocation:
      type: object
      properties:
        id:
          type: integer
        clientProfileId:
          type: integer
        label:
          type: string
        address:
          type: string
        createdAt:
          type: string
          format: date-time
        isPrimary:
          type: boolean
      x-source: prisma/schema.prisma model ClientSavedLocation
...